refactor repo for heroku deployments
