--- api/.flaskenv
@@ -1,2 +1,0 @@
-FLASK_APP=api.py
-FLASK_ENV=development
