+++ app.py
@@ -1,8 +1,8 @@
 import os
 from flask import Flask, request
-from test_data.restaurant_data import RESTAURANTS
+from restaurant_data import RESTAURANTS
 
-app = Flask(__name__, static_folder='../build', static_url_path='/')
+app = Flask(__name__, static_folder='./build', static_url_path='/')
 
 @app.route('/')
 def index():
